Update to fix how bulk dates are transformed
If the date expression includes anything but the text "bulk ", which ASpace includes by default if no value is present, then that value should stay.  e.g. if the date expression is "bulk 1900-1910", the export should be "1900-1910"; but if the date expression is "circa 1900-1910", then the export should be "circa 1900-1910"